Fix some small type errors exposed by clang code model

+++ src/ubuntumirclient/screen.cpp
@@ -312,7 +312,7 @@ void UbuntuScreen::updateMirOutput(const MirOutput *output)
     }
 }
 
-void UbuntuScreen::setAdditionalMirDisplayProperties(float scale, MirFormFactor formFactor, float dpi)
+void UbuntuScreen::setAdditionalMirDisplayProperties(float scale, MirFormFactor formFactor, int dpi)
 {
     if (mDpi != dpi) {
         mDpi = dpi;
